+++ Swift-Cmdline-Utility/main.swift
@@ -1,4 +1,4 @@
+//  Copyright © 2020 Erica Sadun. All rights reserved.
+
 import Foundation
- 
-var german = TransformativeOutputStream.translating(to: "de")
-print("\"I am a jelly donut.\" ✌️", to: &german)
+