+++ Tests/Tests/Tests.swift
@@ -1,0 +1,18 @@
+//  Copyright © 2020 Erica Sadun. All rights reserved.
+
+import XCTest
+import MacUtility
+
+class Tests: XCTestCase {
+
+    override func setUpWithError() throws {
+    }
+
+    override func tearDownWithError() throws {
+    }
+
+    func test() throws {
+        XCTAssertTrue(true)
+    }
+}
+